+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Release notes for the Router Craft CMS plugin.
 
 
 
+## Unreleased
+
+### Added
+- Added a new EntryType filter
+
+
+
 ## 1.0.0-beta.2 - 2019-12-27
 
 ### Added